again

+++ src/edge/status/special_hi.rs
@@ -17,7 +17,7 @@ unsafe fn edge_special_hi_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
         fighter.module_accessor,
         SituationKind(*SITUATION_KIND_NONE),
         *FIGHTER_KINETIC_TYPE_MOTION_CLIFF_MOVE,
-        *GROUND_CORRECT_KIND_KEEP,
+        *GROUND_CORRECT_KIND_KEEP as u32,
         GroundCliffCheckKind(*GROUND_CLIFF_CHECK_KIND_NONE),
         true,
         *FIGHTER_STATUS_WORK_KEEP_FLAG_NONE_FLAG,
@@ -36,7 +36,7 @@ unsafe fn edge_special_hi_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
             *FIGHTER_LOG_MASK_FLAG_ATTACK_KIND_SPECIAL_HI |
             *FIGHTER_LOG_MASK_FLAG_ACTION_CATEGORY_ATTACK |
             *FIGHTER_LOG_MASK_FLAG_ACTION_TRIGGER_ON
-        ) as i64,
+        ) as u64,
         *FIGHTER_STATUS_ATTR_START_TURN as u32,
         *FIGHTER_POWER_UP_ATTACK_BIT_SPECIAL_HI as u32,
         0
